Apiの型ファイルを定義
