+++ src/types/Category.ts
@@ -1,19 +1,16 @@
 // カテゴリーのレスポンス
 // https://developers.newt.so/apis/cdn
 
+import { Raw } from './CommonCdnApi'
+
 export type Category = {
-  _id: '63905e2819c4833e9f78152f'
+  _id: string
   _sys: {
-    raw: {
-      createdAt: '2022-12-07T09:34:32.924Z'
-      updatedAt: '2022-12-07T09:34:32.924Z'
-      firstPublishedAt: '2022-12-07T09:34:32.924Z'
-      publishedAt: '2022-12-07T09:34:32.924Z'
-    }
-    customOrder: 2
-    createdAt: '2022-12-07T09:34:32.924Z'
-    updatedAt: '2022-12-07T09:34:32.924Z'
+    raw: Raw
+    customOrder: number
+    createdAt: string
+    updatedAt: string
   }
-  label: 'テニス'
-  slug: 'tennis'
+  label: string
+  slug: string
 }
